chore: 移除不必要的worker 文件
