+++ package.json
@@ -110,8 +110,8 @@
     "@types/pngjs": "^6.0.3",
     "@types/react-router-dom": "^5.3.2",
     "@types/supercluster": "^5.0.1",
-    "@typescript-eslint/eslint-plugin": "^6.9.0",
-    "@typescript-eslint/parser": "^6.9.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
     "@umijs/fabric": "^2.8.1",
     "@vitejs/plugin-react": "^4.2.1",
     "antd": "^4.12.3",
