--- packages/utils/src/worker-helper/worker-map.ts
@@ -1,10 +1,0 @@
-const WorkerSourceMap = new Map<string, string>();
-
-export function registerWorkerSource(workerName: string, workerSource: string) {
-  WorkerSourceMap.set(workerName, workerSource);
-}
-
-export function getWorkerSource(workerName: string) {
-  const workerSource = WorkerSourceMap.get(workerName);
-  return workerSource;
-}
